\name{soilwaterfun-package}
\alias{soilwaterfun-package}
\alias{soilwaterfun}
\docType{package}
\title{Functions for soil water retention and soil hydraulic conductivity.}
\description{The package implements commonly used functions for soil water retention
(water content = f(matrix potential)) and soil hydraulic conductivity (hydraulic
conductivity K = f(matrix potential), or = f(water content)). The functions for
for soil water retention implemented are: van Genuchten 1980; Brooks & Corey 1964;
Campbell 1974. The functions for soil hydraulic conductivity implemented are: Mualem
1976; Brooks & Corey 1964; Campbell 1974. NB: The package soilwaterptf
contains pedotransfer function for estimating the parameters of these functions
from basic soil properties (valid for european soils).}
\details{
\tabular{ll}{Package: \tab soilwaterfun\cr
Version: \tab 1.0.5\cr
Date: \tab 2014-01-09\cr
Title: \tab Functions for soil water retention and soil hydraulic conductivity.\cr
Author: \tab Julien MOEYS <jules_m78-soiltexture@yahoo.fr>\cr
Maintainer: \tab Julien MOEYS <jules_m78-soiltexture@yahoo.fr>\cr
Depends: \tab R (>= 3.0.2)\cr
Suggests: \tab \cr
License: \tab AGPL-3\cr
URL: \tab http://www.mark.slu.se, http://ckb.slu.se, http://julienmoeys.free.fr\cr}
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}

\keyword{ package }


